adapter,environmentd: standardize on "caught up" nomenclature
The 0dt deployment "hydration check" has morphed into the "caught up
check". Update all variable names and module names to match.
